src/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning
torch==2.5.1
torchaudio==2.1.0
transformers==4.36.0
openai-whisper==20231117
scikit-learn==1.3.2
numpy==1.24.3

# Audio Processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1
python-speech-features==0.6

# Async & Utilities
aiofiles==23.2.1
asyncio==3.4.3
redis==5.0.1
celery==5.3.4

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Environment
python-dotenv==1.0.0